+++ CHANGELOG.md
@@ -8,16 +8,18 @@ Please format the changes as follows:
 + Updates:
 
 ## 1.0.45
++ BugFixes:
+  + Address package restore issue with Build.ps1
 
 ## 1.0.44
 + New:
   + Add support for Windows ARM64 (Internal)
-+ BugFixes: 
++ BugFixes:
   + Fixed build scripts to explicitly point to v142 toolsets regardless of VS version
 + Updates:
   + Setup automated PR builds
   + Add linter for doc link fixes
-  
+
 ## 1.0.43
 + New:
   + On Unix, environment is now scanned for instrumentation methods instead of looking for hardcoded ProductionBreakpoints.