+++ CHANGELOG.md
@@ -8,8 +8,11 @@ Please format the changes as follows:
 + Updates:
 
 ## 1.0.45
++ New:
+  + Add ARM64 environment variable support
 + BugFixes:
   + Refactor duplicate instrumentation method check
+  + Address package restore issue with Build.ps1
 
 ## 1.0.44
 + New: